pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "png2svg"
version = "0.1.0"
description = "Convert PNG images to SVG with advanced options"
readme = "README.md"
readme-content-type = "text/markdown"
authors = [
    { name = "heysarver", email = "your.email@example.com" }
]
license = { text = "MIT License" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]
requires-python = ">=3.7"
dependencies = [
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "scikit-learn>=1.0.0",
    "svgwrite>=1.4.0",
]

[project.urls]
Homepage = "https://github.com/heysarver/png2svg"

[tool.setuptools.packages.find]
where = ["png2svg"]

[tool.setuptools.entry-points.console_scripts]
png2svg = "png2svg.cli:main"
